wip: change menu ui

+++ before.js
@@ -1,7 +1,10 @@
 // this code will be executed before page load
 (function () {
   console.log("before.js executed");
-  chrome.runtime.onStartup.addListener(function() {
-    chrome.storage.local.clear()
-   })
+  if (chrome.runtime.onStartup) {
+    chrome.runtime.onStartup.addListener(function() {
+      chrome.storage.local.clear()
+    })
+  }
+
 })();
